# maize/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product API",
        "description": "API documentation for Product endpoints",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/yield-predictions": {
            "get": {
                "tags": [
                    "YieldPredictions"
                ],
                "summary": "Get list of all yield predictions",
                "description": "Display a listing of the yield predictions.",
                "operationId": "cf8f94ccad91c83592151ec5e3c5af61",
                "responses": {
                    "200": {
                        "description": "List of yield predictions"
                    }
                }
            },
            "post": {
                "tags": [
                    "YieldPredictions"
                ],
                "summary": "Create a new yield prediction",
                "description": "Store a newly created yield prediction in storage.",
                "operationId": "3aac187079df0a7cc60d41ab1122461b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "field_id",
                                    "model_version",
                                    "prediction_date",
                                    "predicted_yield_t_ha"
                                ],
                                "properties": {
                                    "field_id": {
                                        "type": "string"
                                    },
                                    "model_version": {
                                        "type": "string"
                                    },
                                    "prediction_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "predicted_yield_t_ha": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Yield prediction created"
                    }
                }
            }
        },
        "/yield-predictions/{prediction}": {
            "get": {
                "tags": [
                    "YieldPredictions"
                ],
                "summary": "Get a specific yield prediction",
                "description": "Display the specified yield prediction.",
                "operationId": "2c905fdc43497f712f50d9c9e97152f3",
                "parameters": [
                    {
                        "name": "prediction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Yield prediction details"
                    },
                    "404": {
                        "description": "Yield prediction not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "YieldPredictions"
                ],
                "summary": "Update a yield prediction",
                "description": "Update the specified yield prediction in storage.",
                "operationId": "76937df1a1882f3bbcab26a57e1f43d4",
                "parameters": [
                    {
                        "name": "prediction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "field_id": {
                                        "type": "string"
                                    },
                                    "model_version": {
                                        "type": "string"
                                    },
                                    "prediction_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "predicted_yield_t_ha": {
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Yield prediction updated"
                    },
                    "404": {
                        "description": "Yield prediction not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "YieldPredictions"
                ],
                "summary": "Delete a yield prediction",
                "description": "Remove the specified yield prediction from storage.",
                "operationId": "5b2b053a5009b5853eb5189349ead271",
                "parameters": [
                    {
                        "name": "prediction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Yield prediction deleted"
                    },
                    "404": {
                        "description": "Yield prediction not found"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}